moved binder input sequence expansion function to utils
